+++ src/writers/log_writer.rs
@@ -1,3 +1,4 @@
+use crate::deferred_now::DeferredNow;
 use log::Record;
 use std::io;
 
@@ -6,7 +7,7 @@ use std::io;
 /// Boxed instances of `LogWriter` can be used as additional log targets.
 pub trait LogWriter: Sync + Send {
     /// Writes out a log line.
-    fn write(&self, record: &Record) -> io::Result<()>;
+    fn write(&self, now: &mut DeferredNow, record: &Record) -> io::Result<()>;
 
     /// Flushes any buffered records.
     fn flush(&self) -> io::Result<()>;