+++ tests/test_multi_threaded_cleanup.rs
@@ -1,7 +1,9 @@
 #[cfg(feature = "ziplogs")]
 mod d {
     use chrono::Local;
-    use flexi_logger::{Cleanup, Criterion, Duplicate, LogSpecification, Logger, Naming, Record};
+    use flexi_logger::{
+        Cleanup, Criterion, DeferredNow, Duplicate, LogSpecification, Logger, Naming, Record,
+    };
     use glob::glob;
     use log::*;
     use std::io;
@@ -103,11 +105,15 @@ mod d {
         )
     }
 
-    pub fn test_format(w: &mut io::Write, record: &Record) -> io::Result<()> {
+    pub fn test_format(
+        w: &mut io::Write,
+        now: &mut DeferredNow,
+        record: &Record,
+    ) -> io::Result<()> {
         write!(
             w,
             "XXXXX [{}] T[{:?}] {} [{}:{}] {}",
-            Local::now().format("%Y-%m-%d %H:%M:%S%.6f %:z"),
+            now.now().format("%Y-%m-%d %H:%M:%S%.6f %:z"),
             thread::current().name().unwrap_or("<unnamed>"),
             record.level(),
             record.file().unwrap_or("<unnamed>"),