+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2019-05-28
+
+Improve performance for plain stderr logging.
+
+Improve robustnesss for recursive log calls.
+
 ## [0.12.0] - 2019-05-24
 
 Revise handling of record.metadata().target() versus record.module_path().